refactor(api): move swaggers to src
